refactor: move encoder and tokenizer initialization into repective files
